+++ blobxfer/operations/upload.py
@@ -850,7 +850,8 @@ class Uploader(object):
             ase = blobxfer.models.azure.StorageEntity(cont, ed)
             if (self._spec.options.mode ==
                     blobxfer.models.azure.StorageModes.File):
-                dir, _ = blobxfer.operations.azure.file.parse_file_path(name)
+                dir, _, _ = blobxfer.operations.azure.file.parse_file_path(
+                    name)
                 ase.populate_from_file(sa, fp, dir)
             else:
                 ase.populate_from_blob(sa, fp)